+++ src/capturer/simple.rs
@@ -1,6 +1,6 @@
 use windows::Win32::Graphics::{
   Direct3D11::ID3D11Texture2D,
-  Dxgi::{DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTPUT_DESC},
+  Dxgi::{DXGI_OUTDUPL_FRAME_INFO, DXGI_OUTDUPL_POINTER_SHAPE_INFO, DXGI_OUTPUT_DESC},
 };
 
 use crate::utils::Result;
@@ -57,6 +57,13 @@ impl Capturer for SimpleCapturer<'_> {
     self
       .ctx
       .capture_frame(self.buffer.as_mut_ptr(), self.buffer.len(), &self.texture)
+  }
+
+  fn get_pointer(
+    &self,
+    frame_info: &DXGI_OUTDUPL_FRAME_INFO,
+  ) -> Result<(Vec<u8>, u32, DXGI_OUTDUPL_POINTER_SHAPE_INFO)> {
+    self.ctx.get_pointer(&frame_info)
   }
 
   fn safe_capture(&mut self) -> Result<DXGI_OUTDUPL_FRAME_INFO> {